Dockerfile: Chain RUN commands with && to catch problems in all commands
If the commands are chained using semicolons, only the result of the
last command matters when deciding whether the command was run
successfully. This can lead to an unsuccessful "wget" still being
considered successful and the resulting image to be cached. It will not
be rerun in subsequent builds either unless a prior line in the file was
changed. All resulting builds will then miss whatever that RUN command
was supposed to generate.